requirements.txt
mss
pyautogui
Pillow
onnxruntime
numpy
pywin32
colorama
chess
